scripts/requirements.txt
python-hcl2
pyyaml
typer 